write works

+++ libs/stm32fs/stm32fs.h
@@ -75,7 +75,7 @@ struct Stm32fsConfigBlock_t {
 
 struct Stm32fsConfig_t {
     std::vector<Stm32fsConfigBlock_t> Blocks;
-    uint32_t BaseBlockAddress;
+    size_t BaseBlockAddress;
     uint32_t SectorSize; //  2048
     std::function<bool (uint8_t)> fnEraseFlashBlock;
     std::function<bool (uint32_t, uint8_t*, size_t)> fnWriteFlash; // address, data, length
@@ -91,7 +91,7 @@ private:
     uint32_t GetBlockAddress(uint8_t blockNum);
     uint32_t GetBlockFromAddress(uint32_t address);
     bool EraseFlashBlock(uint8_t blockNo);
-    bool isFlashEmpty(size_t address, size_t length, bool reverse, uint32_t *exceptAddr);
+    bool isFlashEmpty(uint32_t address, size_t length, bool reverse, uint32_t *exceptAddr);
     bool isFlashBlockEmpty(uint8_t blockNo);
     bool WriteFlash(uint32_t address, uint8_t *data, size_t length);
     bool ReadFlash(uint32_t address, uint8_t *data, size_t length);
